when running the unit tests we should not get errors because db_url is not defined

+++ c2cgeoportal/tests/functional/__init__.py
@@ -8,27 +8,38 @@ from ConfigParser import ConfigParser
 
 from paste.deploy import appconfig
 
+mapserv_url = None
+db_url = None
+
 curdir = os.path.dirname(os.path.abspath(__file__))
 configfile = os.path.realpath(os.path.join(curdir, 'test.ini'))
 
-cfg = ConfigParser()
-cfg.read(configfile)
-sqlalchemy_url = cfg.get('test', 'sqlalchemy.url')
-mapserv_url = cfg.get('test', 'mapserv.url')
+if os.path.exists(configfile):
+    cfg = ConfigParser()
+    cfg.read(configfile)
+    db_url = cfg.get('test', 'sqlalchemy.url')
+    mapserv_url = cfg.get('test', 'mapserv.url')
 
 def setUpModule():
     import c2cgeoportal
     c2cgeoportal.schema = 'main'
     c2cgeoportal.srid = 21781
 
+    if db_url is None:
+        return
+
     import sqlahelper
     from sqlalchemy import create_engine
-    sqlahelper.add_engine(create_engine(sqlalchemy_url))
+    sqlahelper.add_engine(create_engine(db_url))
 
     from c2cgeoportal.models import Base
     Base.metadata.create_all()
 
 def tearDownModule():
+
+    if db_url is None:
+        return
+
     from c2cgeoportal.models import Base
     Base.metadata.drop_all(checkfirst=True)
